refactor(control-api): extract ack subtree from response_actions/mod.rs
