+++ crates/services/control-api/src/routes/response_actions/ack/mod.rs
@@ -1,0 +1,11 @@
+//! Acknowledgement submission, verification, and persistence subsystem.
+
+mod parse;
+mod persist;
+mod policy_rule_diff;
+mod receipt;
+
+pub(crate) use parse::*;
+pub(crate) use persist::*;
+pub(crate) use policy_rule_diff::*;
+pub(crate) use receipt::*;
