+++ crates/services/control-api/src/routes/response_actions/mod.rs
@@ -44,6 +44,7 @@ policy_rule_diff_validation";
 const ACK_STATUS_ALLOWLIST: &str = "acknowledged, rejected, failed, expired, rolled_back";
 
 mod access;
+mod ack;
 mod create;
 mod delivery;
 mod dto;
@@ -52,6 +53,7 @@ mod fetch;
 mod publish;
 
 pub(crate) use access::*;
+pub(crate) use ack::*;
 pub(crate) use create::*;
 pub(crate) use delivery::*;
 pub(crate) use dto::*;
@@ -272,1219 +274,6 @@ async fn record_agent_ack(
         status: ack.ack_status.to_string(),
         observed_at: ack.observed_at,
     }))
-}
-
-async fn load_ack_action_tenant_id(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    action_id: Uuid,
-) -> Result<Uuid, ApiError> {
-    let row = sqlx::query::query("SELECT tenant_id FROM response_actions WHERE id = $1")
-        .bind(action_id)
-        .fetch_optional(&mut **tx)
-        .await
-        .map_err(ApiError::Database)?
-        .ok_or(ApiError::NotFound)?;
-    row.try_get("tenant_id").map_err(ApiError::Database)
-}
-
-fn parse_ack_submission(input: RecordResponseAckRequest) -> Result<AckSubmission, ApiError> {
-    let ack_status = normalize_ack_status(&input.status)?;
-    let target_kind = ResponseTargetKind::from_str(input.target_kind.trim())?;
-    let target_id = input.target_id.trim();
-    if target_id.is_empty() {
-        return Err(ApiError::BadRequest("target_id is required".to_string()));
-    }
-    validate_ack_field_len("target_id", target_id, ACK_TARGET_ID_MAX_BYTES)?;
-
-    let ack_token = input.ack_token.trim();
-    if ack_token.is_empty() {
-        return Err(ApiError::BadRequest("ack_token is required".to_string()));
-    }
-    validate_ack_field_len("ack_token", ack_token, ACK_TOKEN_MAX_BYTES)?;
-
-    if let Some(message) = input.message.as_deref() {
-        validate_ack_field_len("message", message, ACK_MESSAGE_MAX_BYTES)?;
-    }
-    if let Some(resulting_state) = input.resulting_state.as_deref() {
-        validate_ack_field_len(
-            "resulting_state",
-            resulting_state,
-            ACK_RESULTING_STATE_MAX_BYTES,
-        )?;
-    }
-
-    let now = Utc::now();
-    let observed_at = input.observed_at.unwrap_or(now);
-    validate_ack_observed_at(observed_at, now)?;
-    let raw_payload = input.raw_payload.unwrap_or_else(|| {
-        json!({
-            "status": ack_status,
-            "message": input.message.clone(),
-            "resulting_state": input.resulting_state.clone(),
-        })
-    });
-    validate_ack_raw_payload(&raw_payload)?;
-
-    Ok(AckSubmission {
-        target_kind,
-        target_id: target_id.to_string(),
-        ack_token: ack_token.to_string(),
-        ack_status,
-        observed_at,
-        message: input.message,
-        resulting_state: input.resulting_state,
-        raw_payload,
-    })
-}
-
-fn validate_ack_field_len(field: &str, value: &str, max_len: usize) -> Result<(), ApiError> {
-    if value.len() > max_len {
-        return Err(ApiError::BadRequest(format!(
-            "acknowledgement {field} must be at most {max_len} bytes"
-        )));
-    }
-    Ok(())
-}
-
-fn validate_ack_raw_payload(value: &Value) -> Result<(), ApiError> {
-    let serialized = serde_json::to_vec(value).map_err(|err| {
-        ApiError::BadRequest(format!("acknowledgement raw_payload is invalid: {err}"))
-    })?;
-    if serialized.len() > ACK_RAW_PAYLOAD_MAX_BYTES {
-        return Err(ApiError::BadRequest(format!(
-            "acknowledgement raw_payload must be at most {ACK_RAW_PAYLOAD_MAX_BYTES} bytes"
-        )));
-    }
-    Ok(())
-}
-
-fn validate_ack_observed_at(
-    observed_at: DateTime<Utc>,
-    now: DateTime<Utc>,
-) -> Result<(), ApiError> {
-    if observed_at > now + Duration::seconds(ACK_OBSERVED_AT_FUTURE_SKEW_SECONDS) {
-        return Err(ApiError::BadRequest(format!(
-            "acknowledgement observed_at must not be more than {ACK_OBSERVED_AT_FUTURE_SKEW_SECONDS} seconds in the future"
-        )));
-    }
-    if observed_at < now - Duration::seconds(ACK_OBSERVED_AT_MAX_AGE_SECONDS) {
-        return Err(ApiError::BadRequest(format!(
-            "acknowledgement observed_at must be within the last {ACK_OBSERVED_AT_MAX_AGE_SECONDS} seconds"
-        )));
-    }
-    Ok(())
-}
-
-async fn load_ack_context(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    tenant_id: Uuid,
-    action_id: Uuid,
-    ack: &AckSubmission,
-) -> Result<Option<AckContext>, ApiError> {
-    let action = sqlx::query::query(
-        "SELECT * FROM response_actions WHERE id = $1 AND tenant_id = $2 FOR UPDATE",
-    )
-    .bind(action_id)
-    .bind(tenant_id)
-    .fetch_optional(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?
-    .ok_or(ApiError::NotFound)?;
-    let action = ResponseActionRecord::from_row(action).map_err(ApiError::Database)?;
-    if !action.require_acknowledgement {
-        return Err(ApiError::BadRequest(
-            "acknowledgements are not enabled for this action".to_string(),
-        ));
-    }
-
-    let delivery = sqlx::query::query(
-        r#"SELECT id, status, acknowledgement_deadline, metadata
-           FROM response_action_deliveries
-           WHERE action_id = $1
-             AND tenant_id = $2
-             AND target_kind = $3
-             AND target_id = $4
-           FOR UPDATE"#,
-    )
-    .bind(action.id)
-    .bind(action.tenant_id)
-    .bind(ack.target_kind.as_str())
-    .bind(&ack.target_id)
-    .fetch_optional(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?
-    .ok_or_else(|| {
-        ApiError::BadRequest("acknowledgement target does not match a known delivery".to_string())
-    })?;
-    let delivery_id: Uuid = delivery.try_get("id").map_err(ApiError::Database)?;
-    let delivery_status: String = delivery.try_get("status").map_err(ApiError::Database)?;
-    let acknowledgement_deadline: Option<DateTime<Utc>> = delivery
-        .try_get("acknowledgement_deadline")
-        .map_err(ApiError::Database)?;
-    let delivery_metadata: Value = delivery.try_get("metadata").map_err(ApiError::Database)?;
-    let expected_ack_token = delivery_metadata
-        .get("ack_token")
-        .and_then(Value::as_str)
-        .ok_or_else(|| {
-            ApiError::BadRequest("delivery is not acknowledgement-enabled".to_string())
-        })?;
-    if expected_ack_token != ack.ack_token {
-        return Err(ApiError::Forbidden);
-    }
-    let ack_exists = sqlx::query_scalar::query_scalar::<_, bool>(
-        r#"SELECT EXISTS(
-                   SELECT 1
-                   FROM response_action_acks
-                   WHERE delivery_id = $1
-               )"#,
-    )
-    .bind(delivery_id)
-    .fetch_one(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?;
-    if ack_exists {
-        return Err(ApiError::Conflict(
-            "delivery acknowledgement has already been recorded".to_string(),
-        ));
-    }
-
-    if ensure_ack_window_open(
-        tx,
-        &action,
-        delivery_id,
-        &delivery_status,
-        acknowledgement_deadline,
-        Utc::now(),
-    )
-    .await?
-    {
-        return Ok(None);
-    }
-
-    Ok(Some(AckContext {
-        action,
-        delivery_id,
-    }))
-}
-
-async fn validate_endpoint_ack_signed_receipt(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    context: &AckContext,
-    ack: &AckSubmission,
-) -> Result<(), ApiError> {
-    if context.action.action_type == ResponseActionType::PolicyRuleDiffValidation.as_str() {
-        if ack.ack_status == "acknowledged" {
-            return validate_policy_rule_diff_ack_receipt(tx, context, ack).await;
-        }
-        validate_policy_rule_diff_error_payload_matches_action(
-            ack,
-            &context.action.payload,
-            &ack.target_id,
-        )?;
-        return validate_response_ack_signed_receipt(tx, context, ack).await;
-    }
-
-    if !requires_endpoint_ack_signed_receipt(&context.action, ack) {
-        return Ok(());
-    }
-
-    validate_response_ack_signed_receipt(tx, context, ack).await
-}
-
-async fn validate_response_ack_signed_receipt(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    context: &AckContext,
-    ack: &AckSubmission,
-) -> Result<(), ApiError> {
-    let public_key =
-        load_endpoint_ack_public_key(tx, context.action.tenant_id, &ack.target_id).await?;
-
-    let signed_receipt_value = ack
-        .raw_payload
-        .get("signedReceipt")
-        .or_else(|| ack.raw_payload.get("signed_receipt"))
-        .cloned()
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "endpoint acknowledgement raw_payload must include signedReceipt".to_string(),
-            )
-        })?;
-    let verified = verify_endpoint_decision_signed_receipt_value(
-        signed_receipt_value,
-        &public_key,
-        "endpoint acknowledgement signedReceipt",
-    )?;
-
-    if let Some(local_receipt_hash) = ack
-        .raw_payload
-        .get("localReceiptHash")
-        .and_then(Value::as_str)
-    {
-        if local_receipt_hash != verified.signed_receipt_hash {
-            return Err(ApiError::BadRequest(
-                "endpoint acknowledgement localReceiptHash must match signedReceipt".to_string(),
-            ));
-        }
-    }
-
-    validate_endpoint_ack_receipt_contract(&verified, context, ack)
-}
-
-async fn load_endpoint_ack_public_key(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    tenant_id: Uuid,
-    agent_id: &str,
-) -> Result<PublicKey, ApiError> {
-    let public_key_hex = sqlx::query_scalar::query_scalar::<_, String>(
-        r#"SELECT public_key
-           FROM agents
-           WHERE tenant_id = $1
-             AND agent_id = $2"#,
-    )
-    .bind(tenant_id)
-    .bind(agent_id)
-    .fetch_optional(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?
-    .ok_or_else(|| {
-        ApiError::BadRequest(
-            "endpoint acknowledgement target does not match a registered agent".to_string(),
-        )
-    })?;
-    let public_key = PublicKey::from_hex(public_key_hex.trim()).map_err(|_| {
-        ApiError::BadRequest("endpoint acknowledgement agent public_key is invalid".to_string())
-    })?;
-    Ok(public_key)
-}
-
-fn verify_endpoint_decision_signed_receipt_value(
-    signed_receipt_value: Value,
-    public_key: &PublicKey,
-    label: &str,
-) -> Result<VerifiedEndpointDecisionReceipt, ApiError> {
-    let canonical_signed_receipt = canonicalize_json(&signed_receipt_value).map_err(|err| {
-        ApiError::BadRequest(format!("{label} is not canonicalizable JSON: {err}"))
-    })?;
-    if canonical_signed_receipt.len() > ACK_SIGNED_RECEIPT_MAX_BYTES {
-        return Err(ApiError::BadRequest(format!(
-            "{label} must be no larger than {ACK_SIGNED_RECEIPT_MAX_BYTES} bytes"
-        )));
-    }
-    let signed_receipt: SignedReceipt = serde_json::from_value(signed_receipt_value.clone())
-        .map_err(|err| ApiError::BadRequest(format!("{label} is invalid: {err}")))?;
-    let verification = signed_receipt.verify(&PublicKeySet::new(public_key.clone()));
-    if !verification.valid {
-        let reason = if verification.errors.is_empty() {
-            "unknown verification error".to_string()
-        } else {
-            verification.errors.join("; ")
-        };
-        return Err(ApiError::BadRequest(format!(
-            "{label} failed verification: {reason}"
-        )));
-    }
-
-    let endpoint_decision_value = signed_receipt
-        .receipt
-        .metadata
-        .as_ref()
-        .and_then(|metadata| metadata.get("endpointDecision"))
-        .cloned()
-        .ok_or_else(|| {
-            ApiError::BadRequest(format!(
-                "{label} must include receipt.metadata.endpointDecision"
-            ))
-        })?;
-    let endpoint_decision: EndpointDecisionReceipt =
-        serde_json::from_value(endpoint_decision_value.clone()).map_err(|err| {
-            ApiError::BadRequest(format!("{label} endpointDecision is invalid: {err}"))
-        })?;
-    endpoint_decision.validate().map_err(|err| {
-        ApiError::BadRequest(format!("{label} endpointDecision failed validation: {err}"))
-    })?;
-
-    let canonical_endpoint_decision =
-        canonicalize_json(&endpoint_decision_value).map_err(|err| {
-            ApiError::BadRequest(format!(
-                "{label} endpointDecision is not canonicalizable JSON: {err}"
-            ))
-        })?;
-    let endpoint_decision_hash = sha256(canonical_endpoint_decision.as_bytes());
-    if signed_receipt.receipt.content_hash != endpoint_decision_hash {
-        return Err(ApiError::BadRequest(format!(
-            "{label} content_hash must match receipt.metadata.endpointDecision"
-        )));
-    }
-
-    let signer_public_key = endpoint_decision
-        .signer
-        .signer_public_key
-        .as_deref()
-        .filter(|value| !value.trim().is_empty())
-        .ok_or_else(|| {
-            ApiError::BadRequest(format!(
-                "{label} endpointDecision.signer.signerPublicKey is required"
-            ))
-        })?;
-    let metadata_public_key = PublicKey::from_hex(signer_public_key).map_err(|_| {
-        ApiError::BadRequest(format!(
-            "{label} endpointDecision.signer.signerPublicKey must be a valid Ed25519 public key hex"
-        ))
-    })?;
-    if metadata_public_key.to_hex() != public_key.to_hex() {
-        return Err(ApiError::BadRequest(format!(
-            "{label} endpointDecision.signer.signerPublicKey must match registered agent public_key"
-        )));
-    }
-
-    Ok(VerifiedEndpointDecisionReceipt {
-        signed_receipt_hash: sha256(canonical_signed_receipt.as_bytes()).to_hex_prefixed(),
-        endpoint_decision,
-        endpoint_decision_value,
-    })
-}
-
-async fn validate_policy_rule_diff_ack_receipt(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    context: &AckContext,
-    ack: &AckSubmission,
-) -> Result<(), ApiError> {
-    validate_policy_rule_diff_ack_payload(ack)?;
-    if !matches!(ack.ack_status, "acknowledged") {
-        return Ok(());
-    }
-
-    let payload = policy_rule_diff_ack_payload_value(ack)?;
-    let signed_receipt_value = payload.get("receipt").cloned().ok_or_else(|| {
-        ApiError::BadRequest(
-            "policyRuleDiffValidation acknowledgement must include receipt".to_string(),
-        )
-    })?;
-    validate_policy_rule_diff_ack_matches_action_payload(
-        payload,
-        &context.action.payload,
-        &ack.target_id,
-    )?;
-    let public_key =
-        load_endpoint_ack_public_key(tx, context.action.tenant_id, &ack.target_id).await?;
-    let verified = verify_endpoint_decision_signed_receipt_value(
-        signed_receipt_value,
-        &public_key,
-        "policyRuleDiffValidation receipt",
-    )?;
-
-    if verified.endpoint_decision.receipt_family != EndpointDecisionReceiptFamily::Simulation {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation receipt endpointDecision.receiptFamily must be simulation"
-                .to_string(),
-        ));
-    }
-    if verified.endpoint_decision.decision.rule_id.as_deref()
-        != Some("endpoint.policy_event_impact")
-    {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation receipt endpointDecision.decision.ruleId must be endpoint.policy_event_impact"
-                .to_string(),
-        ));
-    }
-    if verified.endpoint_decision.actor.endpoint_id != ack.target_id {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation receipt endpoint id must match target_id".to_string(),
-        ));
-    }
-
-    let impact_id = payload
-        .pointer("/impact/impactId")
-        .and_then(Value::as_str)
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation impact must include impactId".to_string(),
-            )
-        })?;
-    let receipt_impact_id = verified
-        .endpoint_decision
-        .decision
-        .finding_id
-        .as_deref()
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation receipt endpointDecision.decision.findingId is required"
-                    .to_string(),
-            )
-        })?;
-    if impact_id != receipt_impact_id {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation impactId must match the signed receipt findingId".to_string(),
-        ));
-    }
-    validate_policy_rule_diff_ack_impact_evidence(payload, &verified)
-}
-
-fn validate_policy_rule_diff_ack_matches_action_payload(
-    payload: &Value,
-    action_payload: &Value,
-    target_id: &str,
-) -> Result<(), ApiError> {
-    if let Some(expected_proposal_id) = action_payload.get("proposalId").and_then(Value::as_str) {
-        let actual_proposal_id = payload
-            .get("proposalId")
-            .and_then(Value::as_str)
-            .ok_or_else(|| {
-                ApiError::BadRequest(
-                    "policyRuleDiffValidation acknowledgement must include proposalId".to_string(),
-                )
-            })?;
-        if actual_proposal_id != expected_proposal_id {
-            return Err(ApiError::BadRequest(
-                "policyRuleDiffValidation proposalId does not match the dispatched response action"
-                    .to_string(),
-            ));
-        }
-    }
-
-    if let Some(expected_plan_sha256) = action_payload
-        .get("validationPlanSha256")
-        .and_then(Value::as_str)
-    {
-        let actual_plan_sha256 = payload
-            .get("validationPlanSha256")
-            .and_then(Value::as_str)
-            .ok_or_else(|| {
-                ApiError::BadRequest(
-                    "policyRuleDiffValidation acknowledgement must include validationPlanSha256"
-                        .to_string(),
-                )
-            })?;
-        if actual_plan_sha256 != expected_plan_sha256 {
-            return Err(ApiError::BadRequest(
-                "policyRuleDiffValidation validationPlanSha256 does not match the dispatched response action"
-                    .to_string(),
-            ));
-        }
-    }
-
-    let actual_endpoint_agent_id = payload
-        .get("endpointAgentId")
-        .and_then(Value::as_str)
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation acknowledgement must include endpointAgentId".to_string(),
-            )
-        })?;
-    if actual_endpoint_agent_id != target_id {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation endpointAgentId does not match target_id".to_string(),
-        ));
-    }
-    if let Some(expected_endpoint_agent_id) = action_payload
-        .get("endpointAgentId")
-        .and_then(Value::as_str)
-    {
-        if actual_endpoint_agent_id != expected_endpoint_agent_id {
-            return Err(ApiError::BadRequest(
-                "policyRuleDiffValidation endpointAgentId does not match the dispatched response action"
-                    .to_string(),
-            ));
-        }
-    }
-
-    if let Some(expected_request) = action_payload.get("request") {
-        let actual_request = payload.get("request").ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation acknowledgement must include request".to_string(),
-            )
-        })?;
-        let expected_request = canonicalize_json(expected_request).map_err(|err| {
-            ApiError::BadRequest(format!(
-                "policyRuleDiffValidation dispatched request is not canonicalizable JSON: {err}"
-            ))
-        })?;
-        let actual_request = canonicalize_json(actual_request).map_err(|err| {
-            ApiError::BadRequest(format!(
-                "policyRuleDiffValidation acknowledgement request is not canonicalizable JSON: {err}"
-            ))
-        })?;
-        if actual_request != expected_request {
-            return Err(ApiError::BadRequest(
-                "policyRuleDiffValidation request does not match the dispatched response action"
-                    .to_string(),
-            ));
-        }
-    }
-
-    if let Some(expected_receipt) = action_payload.get("expectedReceipt") {
-        let actual_expected_receipt = payload.get("expectedReceipt").ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation acknowledgement must include expectedReceipt".to_string(),
-            )
-        })?;
-        let expected_canonical = canonicalize_json(expected_receipt).map_err(|err| {
-            ApiError::BadRequest(format!(
-                "policyRuleDiffValidation dispatched expectedReceipt is not canonicalizable JSON: {err}"
-            ))
-        })?;
-        let actual_canonical = canonicalize_json(actual_expected_receipt).map_err(|err| {
-            ApiError::BadRequest(format!(
-                "policyRuleDiffValidation acknowledgement expectedReceipt is not canonicalizable JSON: {err}"
-            ))
-        })?;
-        if actual_canonical != expected_canonical {
-            return Err(ApiError::BadRequest(
-                "policyRuleDiffValidation expectedReceipt does not match the dispatched response action"
-                    .to_string(),
-            ));
-        }
-        validate_policy_rule_diff_ack_expected_policy_identity(payload, expected_receipt)?;
-    }
-
-    Ok(())
-}
-
-fn validate_policy_rule_diff_ack_expected_policy_identity(
-    payload: &Value,
-    expected_receipt: &Value,
-) -> Result<(), ApiError> {
-    let Some(expected_policy_hash) = expected_receipt
-        .get("proposedPolicyHash")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    else {
-        return Ok(());
-    };
-    let Some(expected_policy_epoch) = expected_receipt
-        .get("proposedPolicyEpoch")
-        .and_then(Value::as_u64)
-    else {
-        return Ok(());
-    };
-
-    let actual_expected_receipt = payload.get("expectedReceipt").ok_or_else(|| {
-        ApiError::BadRequest(
-            "policyRuleDiffValidation acknowledgement must include expectedReceipt".to_string(),
-        )
-    })?;
-    let actual_expected_policy_hash = actual_expected_receipt
-        .get("proposedPolicyHash")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation expectedReceipt must include proposedPolicyHash"
-                    .to_string(),
-            )
-        })?;
-    if actual_expected_policy_hash != expected_policy_hash {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation expectedReceipt proposedPolicyHash does not match the dispatched response action"
-                .to_string(),
-        ));
-    }
-    let actual_expected_policy_epoch = actual_expected_receipt
-        .get("proposedPolicyEpoch")
-        .and_then(Value::as_u64)
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation expectedReceipt must include proposedPolicyEpoch"
-                    .to_string(),
-            )
-        })?;
-    if actual_expected_policy_epoch != expected_policy_epoch {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation expectedReceipt proposedPolicyEpoch does not match the dispatched response action"
-                .to_string(),
-        ));
-    }
-
-    let impact = payload.get("impact").ok_or_else(|| {
-        ApiError::BadRequest(
-            "policyRuleDiffValidation acknowledgement must include impact".to_string(),
-        )
-    })?;
-    let impact_policy_hash = impact
-        .get("proposedPolicyHash")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation impact must include proposedPolicyHash".to_string(),
-            )
-        })?;
-    if impact_policy_hash != expected_policy_hash {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation impact proposedPolicyHash does not match expectedReceipt"
-                .to_string(),
-        ));
-    }
-    let impact_policy_epoch = impact
-        .get("proposedPolicyEpoch")
-        .and_then(Value::as_u64)
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidation impact must include proposedPolicyEpoch".to_string(),
-            )
-        })?;
-    if impact_policy_epoch != expected_policy_epoch {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation impact proposedPolicyEpoch does not match expectedReceipt"
-                .to_string(),
-        ));
-    }
-
-    Ok(())
-}
-
-fn policy_rule_diff_error_payload_value(ack: &AckSubmission) -> Result<&Value, ApiError> {
-    ack.raw_payload
-        .get("policyRuleDiffValidationError")
-        .or_else(|| ack.raw_payload.get("policy_rule_diff_validation_error"))
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "acknowledgement raw_payload must include policyRuleDiffValidationError"
-                    .to_string(),
-            )
-        })
-}
-
-fn validate_policy_rule_diff_error_payload_matches_action(
-    ack: &AckSubmission,
-    action_payload: &Value,
-    target_id: &str,
-) -> Result<(), ApiError> {
-    let payload = policy_rule_diff_error_payload_value(ack)?;
-    if !payload.is_object() {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidationError must be an object".to_string(),
-        ));
-    }
-
-    require_policy_rule_diff_error_str_matches_action(
-        payload,
-        action_payload,
-        "proposalId",
-        "proposalId",
-    )?;
-    require_policy_rule_diff_error_str_matches_action(
-        payload,
-        action_payload,
-        "validationPlanSha256",
-        "validationPlanSha256",
-    )?;
-    let endpoint_agent_id = required_policy_rule_diff_error_str(payload, "endpointAgentId")?;
-    if endpoint_agent_id != target_id {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidationError endpointAgentId does not match target_id".to_string(),
-        ));
-    }
-    let expected_endpoint_agent_id =
-        required_policy_rule_diff_action_str(action_payload, "endpointAgentId")?;
-    if endpoint_agent_id != expected_endpoint_agent_id {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidationError endpointAgentId does not match the dispatched response action"
-                .to_string(),
-        ));
-    }
-
-    let expected_request = action_payload.get("request").ok_or_else(|| {
-        ApiError::BadRequest(
-            "policyRuleDiffValidation dispatched response action must include request".to_string(),
-        )
-    })?;
-    let actual_request = payload.get("request").ok_or_else(|| {
-        ApiError::BadRequest("policyRuleDiffValidationError must include request".to_string())
-    })?;
-    let expected_request = canonicalize_json(expected_request).map_err(|err| {
-        ApiError::BadRequest(format!(
-            "policyRuleDiffValidation dispatched request is not canonicalizable JSON: {err}"
-        ))
-    })?;
-    let actual_request = canonicalize_json(actual_request).map_err(|err| {
-        ApiError::BadRequest(format!(
-            "policyRuleDiffValidationError request is not canonicalizable JSON: {err}"
-        ))
-    })?;
-    if actual_request != expected_request {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidationError request does not match the dispatched response action"
-                .to_string(),
-        ));
-    }
-
-    let message = payload
-        .get("message")
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "policyRuleDiffValidationError must include a non-empty message".to_string(),
-            )
-        })?;
-    if let Some(ack_message) = ack
-        .message
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    {
-        if message != ack_message {
-            return Err(ApiError::BadRequest(
-                "policyRuleDiffValidationError message must match acknowledgement message"
-                    .to_string(),
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-fn require_policy_rule_diff_error_str_matches_action(
-    payload: &Value,
-    action_payload: &Value,
-    payload_key: &str,
-    action_key: &str,
-) -> Result<(), ApiError> {
-    let actual = required_policy_rule_diff_error_str(payload, payload_key)?;
-    let expected = required_policy_rule_diff_action_str(action_payload, action_key)?;
-    if actual != expected {
-        return Err(ApiError::BadRequest(format!(
-            "policyRuleDiffValidationError {payload_key} does not match the dispatched response action"
-        )));
-    }
-    Ok(())
-}
-
-fn required_policy_rule_diff_error_str<'a>(
-    payload: &'a Value,
-    key: &str,
-) -> Result<&'a str, ApiError> {
-    payload
-        .get(key)
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| {
-            ApiError::BadRequest(format!("policyRuleDiffValidationError must include {key}"))
-        })
-}
-
-fn required_policy_rule_diff_action_str<'a>(
-    payload: &'a Value,
-    key: &str,
-) -> Result<&'a str, ApiError> {
-    payload
-        .get(key)
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| {
-            ApiError::BadRequest(format!(
-                "policyRuleDiffValidation dispatched response action must include {key}"
-            ))
-        })
-}
-
-fn validate_policy_rule_diff_ack_impact_evidence(
-    payload: &Value,
-    verified: &VerifiedEndpointDecisionReceipt,
-) -> Result<(), ApiError> {
-    let impact = payload.get("impact").ok_or_else(|| {
-        ApiError::BadRequest(
-            "policyRuleDiffValidation acknowledgement must include impact".to_string(),
-        )
-    })?;
-    for key in [
-        "impactId",
-        "eventStreamHash",
-        "currentResultHash",
-        "proposedResultHash",
-        "impactHash",
-        "proposedPolicyHash",
-        "proposedPolicyEpoch",
-        "eventCount",
-        "changedCount",
-        "allowToBlockCount",
-        "trackPosture",
-    ] {
-        let raw_value = policy_rule_diff_ack_impact_evidence_value(impact, key)?;
-        require_policy_rule_diff_ack_receipt_evidence_hash(
-            &verified.endpoint_decision_value,
-            key,
-            &raw_value,
-        )?;
-    }
-    Ok(())
-}
-
-fn policy_rule_diff_ack_impact_evidence_value(
-    impact: &Value,
-    key: &str,
-) -> Result<String, ApiError> {
-    let value = match key {
-        "proposedPolicyHash" => impact
-            .pointer("/proposedPolicy/policyHash")
-            .and_then(Value::as_str)
-            .map(str::to_string),
-        "proposedPolicyEpoch" => impact
-            .pointer("/proposedPolicy/policyEpoch")
-            .and_then(Value::as_u64)
-            .map(|value| value.to_string()),
-        "eventCount" | "changedCount" | "allowToBlockCount" => impact
-            .get(key)
-            .and_then(Value::as_u64)
-            .map(|value| value.to_string()),
-        "trackPosture" => impact
-            .get(key)
-            .and_then(Value::as_bool)
-            .map(|value| value.to_string()),
-        _ => impact.get(key).and_then(Value::as_str).map(str::to_string),
-    };
-    value.ok_or_else(|| {
-        ApiError::BadRequest(format!(
-            "policyRuleDiffValidation impact must include {key}"
-        ))
-    })
-}
-
-fn require_policy_rule_diff_ack_receipt_evidence_hash(
-    endpoint_decision: &Value,
-    key: &str,
-    raw_value: &str,
-) -> Result<(), ApiError> {
-    let expected_hash = sha256(raw_value.as_bytes()).to_hex_prefixed();
-    let evidence_hash = endpoint_decision
-        .get("evidence")
-        .and_then(Value::as_array)
-        .and_then(|items| {
-            items.iter().find_map(|item| {
-                if item.get("key").and_then(Value::as_str) == Some(key) {
-                    item.get("valueHash").and_then(Value::as_str)
-                } else {
-                    None
-                }
-            })
-        })
-        .ok_or_else(|| {
-            ApiError::BadRequest(format!(
-                "policyRuleDiffValidation receipt evidence is missing {key}"
-            ))
-        })?;
-    if evidence_hash != expected_hash {
-        return Err(ApiError::BadRequest(format!(
-            "policyRuleDiffValidation impact {key} does not match signed receipt evidence"
-        )));
-    }
-    Ok(())
-}
-
-fn policy_rule_diff_ack_payload_value(ack: &AckSubmission) -> Result<&Value, ApiError> {
-    ack.raw_payload
-        .get("policyRuleDiffValidation")
-        .or_else(|| ack.raw_payload.get("policy_rule_diff_validation"))
-        .ok_or_else(|| {
-            ApiError::BadRequest(
-                "acknowledgement raw_payload must include policyRuleDiffValidation".to_string(),
-            )
-        })
-}
-
-fn requires_endpoint_ack_signed_receipt(
-    action: &ResponseActionRecord,
-    ack: &AckSubmission,
-) -> bool {
-    action.target.kind.as_str() == "endpoint"
-        && matches!(
-            ack.ack_status,
-            "acknowledged" | "rolled_back" | "failed" | "rejected" | "expired"
-        )
-}
-
-fn validate_policy_rule_diff_ack_payload(ack: &AckSubmission) -> Result<(), ApiError> {
-    if !matches!(ack.ack_status, "acknowledged") {
-        return Ok(());
-    }
-
-    let payload = policy_rule_diff_ack_payload_value(ack)?;
-
-    if payload.get("receipt").is_none() {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation acknowledgement must include receipt".to_string(),
-        ));
-    }
-    if payload.get("impact").is_none() {
-        return Err(ApiError::BadRequest(
-            "policyRuleDiffValidation acknowledgement must include impact".to_string(),
-        ));
-    }
-
-    Ok(())
-}
-
-fn validate_endpoint_ack_receipt_contract(
-    verified: &VerifiedEndpointDecisionReceipt,
-    context: &AckContext,
-    ack: &AckSubmission,
-) -> Result<(), ApiError> {
-    let endpoint_decision = &verified.endpoint_decision;
-    if endpoint_decision.receipt_family != EndpointDecisionReceiptFamily::ResponseAcknowledgement {
-        return Err(ApiError::BadRequest(
-            "endpoint acknowledgement signedReceipt must be a response_acknowledgement receipt"
-                .to_string(),
-        ));
-    }
-    if endpoint_decision.actor.endpoint_id != ack.target_id {
-        return Err(ApiError::BadRequest(
-            "endpoint acknowledgement signedReceipt endpoint id must match target_id".to_string(),
-        ));
-    }
-
-    let endpoint_decision_value = &verified.endpoint_decision_value;
-    require_endpoint_ack_receipt_evidence_hash(
-        endpoint_decision_value,
-        "controlResponseActionId",
-        &context.action.id.to_string(),
-    )?;
-    require_endpoint_ack_receipt_evidence_hash(
-        endpoint_decision_value,
-        "controlTargetId",
-        &ack.target_id,
-    )?;
-    require_endpoint_ack_receipt_evidence_hash(
-        endpoint_decision_value,
-        "controlAckStatus",
-        ack.ack_status,
-    )?;
-    if let Some(resulting_state) = ack.resulting_state.as_deref() {
-        require_endpoint_ack_receipt_evidence_hash(
-            endpoint_decision_value,
-            "controlResultingState",
-            resulting_state,
-        )?;
-    }
-    let acknowledged_status = ack
-        .resulting_state
-        .as_deref()
-        .map(endpoint_acknowledged_status_from_resulting_state)
-        .unwrap_or(ack.ack_status);
-    require_endpoint_ack_receipt_evidence_hash(
-        endpoint_decision_value,
-        "acknowledgedStatus",
-        acknowledged_status,
-    )?;
-    require_endpoint_ack_receipt_evidence_hash(
-        endpoint_decision_value,
-        "controlAckTokenHash",
-        &sha256(ack.ack_token.as_bytes()).to_hex_prefixed(),
-    )?;
-    if let Some(local_execution_id) = ack
-        .raw_payload
-        .get("localExecutionId")
-        .and_then(Value::as_str)
-    {
-        require_endpoint_ack_receipt_evidence_hash(
-            endpoint_decision_value,
-            "executionId",
-            local_execution_id,
-        )?;
-    }
-    Ok(())
-}
-
-fn endpoint_acknowledged_status_from_resulting_state(resulting_state: &str) -> &str {
-    resulting_state
-        .rsplit_once(':')
-        .map(|(_, status)| status)
-        .unwrap_or(resulting_state)
-}
-
-fn require_endpoint_ack_receipt_evidence_hash(
-    endpoint_decision: &Value,
-    key: &str,
-    raw_value: &str,
-) -> Result<(), ApiError> {
-    let expected_hash = sha256(raw_value.as_bytes()).to_hex_prefixed();
-    let evidence_hash = endpoint_decision
-        .get("evidence")
-        .and_then(Value::as_array)
-        .and_then(|items| {
-            items.iter().find_map(|item| {
-                if item.get("key").and_then(Value::as_str) == Some(key) {
-                    item.get("valueHash").and_then(Value::as_str)
-                } else {
-                    None
-                }
-            })
-        })
-        .ok_or_else(|| {
-            ApiError::BadRequest(format!(
-                "endpoint acknowledgement signedReceipt evidence is missing {key}"
-            ))
-        })?;
-    if evidence_hash != expected_hash {
-        return Err(ApiError::BadRequest(format!(
-            "endpoint acknowledgement signedReceipt evidence {key} does not match acknowledgement"
-        )));
-    }
-    Ok(())
-}
-
-async fn ensure_ack_window_open(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    action: &ResponseActionRecord,
-    delivery_id: Uuid,
-    delivery_status: &str,
-    acknowledgement_deadline: Option<DateTime<Utc>>,
-    now: DateTime<Utc>,
-) -> Result<bool, ApiError> {
-    if action.status != "published" {
-        return Err(ApiError::Conflict(format!(
-            "action status '{}' cannot accept acknowledgements",
-            action.status
-        )));
-    }
-
-    if delivery_status != "published" {
-        return Err(ApiError::Conflict(format!(
-            "delivery status '{}' cannot accept acknowledgements",
-            delivery_status
-        )));
-    }
-
-    let window_expired = action
-        .expires_at
-        .is_some_and(|expires_at| expires_at <= now)
-        || acknowledgement_deadline.is_some_and(|deadline| deadline <= now);
-    if !window_expired {
-        return Ok(false);
-    }
-
-    expire_ack_window(tx, action, delivery_id).await?;
-    Ok(true)
-}
-
-async fn expire_ack_window(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    action: &ResponseActionRecord,
-    delivery_id: Uuid,
-) -> Result<(), ApiError> {
-    sqlx::query::query(
-        r#"UPDATE response_actions
-           SET status = 'expired',
-               updated_at = now()
-           WHERE id = $1
-             AND tenant_id = $2
-             AND status = 'published'"#,
-    )
-    .bind(action.id)
-    .bind(action.tenant_id)
-    .execute(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?;
-
-    sqlx::query::query(
-        r#"UPDATE response_action_deliveries
-           SET status = 'expired',
-               updated_at = now(),
-               last_error = COALESCE(last_error, 'acknowledgement window expired')
-           WHERE id = $1
-             AND action_id = $2
-             AND tenant_id = $3
-             AND status = 'published'"#,
-    )
-    .bind(delivery_id)
-    .bind(action.id)
-    .bind(action.tenant_id)
-    .execute(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?;
-
-    Ok(())
-}
-
-async fn persist_ack_submission(
-    tx: &mut Transaction<'_, sqlx_postgres::Postgres>,
-    context: &AckContext,
-    ack: &AckSubmission,
-) -> Result<(), ApiError> {
-    sqlx::query::query(
-        r#"INSERT INTO response_action_acks (
-               delivery_id,
-               action_id,
-               tenant_id,
-               target_kind,
-               target_id,
-                observed_at,
-               status,
-               message,
-               resulting_state,
-               raw_payload
-           )
-           VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)"#,
-    )
-    .bind(context.delivery_id)
-    .bind(context.action.id)
-    .bind(context.action.tenant_id)
-    .bind(ack.target_kind.as_str())
-    .bind(&ack.target_id)
-    .bind(ack.observed_at)
-    .bind(ack.ack_status)
-    .bind(ack.message.as_deref())
-    .bind(ack.resulting_state.as_deref())
-    .bind(&ack.raw_payload)
-    .execute(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?;
-
-    sqlx::query::query(
-        r#"UPDATE response_action_deliveries
-           SET status = $4,
-               acknowledged_at = $3,
-               updated_at = now(),
-               last_error = CASE WHEN $4 IN ('failed', 'rejected', 'expired') THEN COALESCE($5, last_error) ELSE last_error END
-           WHERE id = $6
-             AND action_id = $1
-             AND tenant_id = $2"#,
-    )
-    .bind(context.action.id)
-    .bind(context.action.tenant_id)
-    .bind(ack.observed_at)
-    .bind(ack.ack_status)
-    .bind(ack.message.as_deref())
-    .bind(context.delivery_id)
-    .execute(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?;
-
-    sqlx::query::query(
-        r#"UPDATE response_actions
-           SET status = $3,
-               updated_at = now()
-           WHERE id = $1 AND tenant_id = $2"#,
-    )
-    .bind(context.action.id)
-    .bind(context.action.tenant_id)
-    .bind(ack.ack_status)
-    .execute(&mut **tx)
-    .await
-    .map_err(ApiError::Database)?;
-
-    Ok(())
-}
-
-fn normalize_ack_status(status: &str) -> Result<&'static str, ApiError> {
-    let status = status.trim();
-    validate_control_discriminator_len("status", status, ACK_STATUS_MAX_BYTES)?;
-    match status {
-        "acknowledged" => Ok("acknowledged"),
-        "rejected" => Ok("rejected"),
-        "failed" => Ok("failed"),
-        "expired" => Ok("expired"),
-        "rolled_back" => Ok("rolled_back"),
-        _ => Err(ApiError::BadRequest(format!(
-            "unsupported ack status; allowed values: {ACK_STATUS_ALLOWLIST}"
-        ))),
-    }
 }
 
 #[cfg(test)]
